refactor: integrate control buttons on MainPage to NavigationBar
