+++ src/pages/MainPage.tsx
@@ -1,12 +1,10 @@
 import React, { useState, useRef, useEffect } from 'react';
 
-import { useNavigate } from 'react-router-dom';
 import { styled } from 'styled-components';
 
 import {
   IconLabel,
   DotPagination,
-  IconButton,
   CardView,
   EmptyView,
   ItemInputModal,
@@ -45,18 +43,8 @@ const MainViewSection = styled.div`
   }
 `;
 
-const ControlSection = styled.div`
-  grid-row: 3 / 4;
-  grid-column: 2 / 3;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 10px;
-`;
-
 export default function MainPage() {
   const [theme] = useTheme();
-  const navigate = useNavigate();
   const pageList = [
     ['WINE', 'Wine'],
     ['WHISKEY', 'Whiskey'],
@@ -95,12 +83,6 @@ export default function MainPage() {
         });
       }
     }, 100);
-  };
-
-  const handleIconClick = (
-    e: React.MouseEvent<HTMLButtonElement, MouseEvent>,
-  ) => {
-    navigate(`/${e.currentTarget.id}`);
   };
 
   useEffect(() => {
@@ -157,27 +139,6 @@ export default function MainPage() {
       {itemInputModalOpen && (
         <ItemInputModal setModalOpen={setItemInputModalOpen} theme={theme} />
       )}
-
-      <ControlSection>
-        <IconButton
-          icon="SETTING"
-          onClick={handleIconClick}
-          size={20}
-          buttonColor="primary"
-        />
-        <IconButton
-          icon="PLUS"
-          onClick={() => setItemInputModalOpen(true)}
-          size={20}
-          buttonColor="primary"
-        />
-        <IconButton
-          icon="USER"
-          onClick={handleIconClick}
-          size={20}
-          buttonColor="primary"
-        />
-      </ControlSection>
     </Container>
   );
 }
